pyproject.toml
[project]
name = "projet-10"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "faiss-cpu>=1.13.2",
    "langchain-core>=1.2.18",
    "langchain-mistralai>=1.1.1",
    "langchain-text-splitters>=1.1.1",
    "matplotlib>=3.10.8",
    "mistralai<2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "pydantic-ai>=1.67.0",
    "pydantic-ai-slim[logfire,mistral]>=1.67.0",
    "pymupdf>=1.27.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "ragas==0.4.3",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
